+++ packages/library/src/components/DaisyAvatarGroup/stories/Playground.stories.ts
@@ -1,0 +1,16 @@
+import {
+    getMeta,
+    type DaisyAvatarGroupMeta,
+    type DaisyAvatarGroupStory
+} from './DaisyAvatarGroupMeta'
+const meta = getMeta()
+export default {
+    ...meta,
+    title: 'Components/DaisyAvatarGroup'
+} as DaisyAvatarGroupMeta
+
+export const Playground: DaisyAvatarGroupStory = {
+    args: {
+        mask: 'circle'
+    }
+}