+++ packages/library/src/components/DaisyBadge/stories/Variations.stories.ts
@@ -18,14 +18,6 @@ export default {
         docs: {
             source: false
         }
-    },
-    argTypes: {
-        ...srcArgTypes,
-        label: {
-            table: {
-                disable: true
-            }
-        }
     }
 } as DaisyBadgeMeta
 
@@ -33,9 +25,7 @@ export const Colors: DaisyBadgeStory = {
     argTypes: {
         ...srcArgTypes,
         color: {
-            table: {
-                disable: true
-            }
+            control: false
         }
     },
     render: (args) => ({
@@ -49,7 +39,8 @@ export const Colors: DaisyBadgeStory = {
         template: `
             <div class="grid grid-cols-5 gap-4">
                 <div v-for="color in daisyBadgeColors" :key="color" class="flex flex-col items-center">
-                    <DaisyBadge :color="color" v-bind="args">{{ color?.toUpperCase() ?? 'UNDEFINED' }}</DaisyBadge>
+                    {{ color?.toUpperCase() ?? 'UNDEFINED' }}
+                    <DaisyBadge :color="color" v-bind="args">{{args.label}}</DaisyBadge>
                 </div>
             </div>
         `
@@ -60,9 +51,7 @@ export const Sizes: DaisyBadgeStory = {
     argTypes: {
         ...srcArgTypes,
         size: {
-            table: {
-                disable: true
-            }
+            control: false
         }
     },
     render: (args) => ({
@@ -76,7 +65,8 @@ export const Sizes: DaisyBadgeStory = {
         template: `
             <div class="grid grid-cols-5 gap-4">
                 <div v-for="size in daisyBadgeSizes" :key="size" class="flex flex-col items-center">
-                    <DaisyBadge :size="size" v-bind="args">{{ size?.toUpperCase() ?? 'UNDEFINED' }}</DaisyBadge>
+                    {{ size?.toUpperCase() ?? 'UNDEFINED' }}
+                    <DaisyBadge :size="size" v-bind="args">{{args.label}}</DaisyBadge>
                 </div>
             </div>
         `
@@ -86,16 +76,8 @@ export const Sizes: DaisyBadgeStory = {
 export const Variants: DaisyBadgeStory = {
     argTypes: {
         ...srcArgTypes,
-        ...Object.keys(srcArgTypes).reduce((result, key) => {
-            if (key !== 'color' && key !== 'size') {
-                result[key] = {
-                    table: {
-                        disable: true
-                    }
-                }
-            }
-            return result
-        }, {})
+        label: { control: false },
+        outline: { control: false }
     },
     render: (args) => ({
         components: { DaisyBadge },
@@ -103,12 +85,17 @@ export const Variants: DaisyBadgeStory = {
             return { args }
         },
         template: `
-            <div style="width: 896px;" class=" m-4 flex flex-col items-center">
-                <DaisyBadge outline v-bind="args">24</DaisyBadge>
+            <div class="grid grid-cols-2 gap-4">
+                <div class=" flex flex-col items-center">
+                    OUTLINE
+                    <DaisyBadge outline v-bind="args">24</DaisyBadge>
+                </div>
+                <div class=" flex flex-col items-center">
+                    NO CONTENT
+                    <DaisyBadge v-bind="args"/>
+                </div>
             </div>
-            <div style="width: 896px;" class=" flex flex-col items-center">
-                <DaisyBadge v-bind="args"/>
-            </div>
+            
         `
     })
 }