+++ packages/library/src/components/DaisyKbd/stories/Variations.stories.ts
@@ -1,11 +1,6 @@
-import {
-    getMeta,
-    type DaisyKbdMeta,
-    type DaisyKbdStory
-} from './DaisyKbdMeta'
+import { getMeta, type DaisyKbdMeta, type DaisyKbdStory } from './DaisyKbdMeta'
 import DaisyKbd from '../DaisyKbd.vue'
 import { sizes } from '../../../globals'
-
 
 const meta = getMeta()
 const srcArgTypes = meta.argTypes
@@ -18,14 +13,6 @@ export default {
         docs: {
             source: false
         }
-    },
-    argTypes: {
-        ...srcArgTypes,
-        label: {
-            table: {
-                disable: true
-            }
-        }
     }
 } as DaisyKbdMeta
 
@@ -33,9 +20,7 @@ export const Sizes: DaisyKbdStory = {
     argTypes: {
         ...srcArgTypes,
         size: {
-            table: {
-                disable: true
-            }
+            control: false
         }
     },
     render: (args) => ({
@@ -49,7 +34,8 @@ export const Sizes: DaisyKbdStory = {
         template: `
             <div class="grid grid-cols-5 gap-4">
                 <div v-for="size in sizes" :key="size" class="flex flex-col items-center">
-                    <DaisyKbd :size="size" v-bind="args">{{ size?.toUpperCase() ?? 'UNDEFINED' }}</DaisyKbd>
+                    {{ size?.toUpperCase() ?? 'UNDEFINED' }}
+                    <DaisyKbd :size="size" v-bind="args">{{ args.label }}</DaisyKbd>
                 </div>
             </div>
         `