+++ packages/library/src/components/components.ts
@@ -7,6 +7,8 @@ import DaisyBrowser from './DaisyBrowser/DaisyBrowser.vue'
 import DaisyKbd from './DaisyKbd/DaisyKbd.vue'
 import DaisyAccordion from './DaisyAccordion/DaisyAccordion.vue'
 import DaisyAccordionPanel from './DaisyAccordionPanel/DaisyAccordionPanel.vue'
+import DaisyAvatar from './DaisyAvatar/DaisyAvatar.vue'
+import DaisyAvatarGroup from './DaisyAvatarGroup/DaisyAvatarGroup.vue'
 // import - do not remove this line, used for hygen generations
 
 // ignore removal of traling comma by prettier so hygen can add new imports
@@ -21,5 +23,7 @@ export {
   DaisyKbd,
   DaisyAccordion,
   DaisyAccordionPanel,
+  DaisyAvatar,
+  DaisyAvatarGroup,
   // export - do not remove this line, used for hygen generations
 }